fix llamadas al constructor de fachada, avance caso de uso 3

+++ Obligatorio/include/fachada.h
@@ -20,7 +20,7 @@ class Fachada{
         /** Requerimientos **/
         void registrarLibro(Libro *);
         void listarLibros();
-        void listarLibro(long int);
+        void listarLibro(int);
         void registrarAutor(Autor);
         void listarAutores();
         float calcularMontoTotal();
